Added new API for AtomicFile - NoSync to address power regression
A power regression on high data rate networks such as LTE
is reported and was traced to the collection of Data Stats.
This regression occurs due to repeated 'writes' to the
Atomic files as a result of LIMIT_GLOBAL_ALERT feature.
'fsync' (FileUtils.sync) after every write to the atomic files
is a Power/CPU consuming operation. Hence added a new API
'finishWriteNoSync' to address the power consumption/ CPU
hogging issue.

CRs-Fixed: 336273

Change-Id: I56817102a67c0d75a252fb362a814802adaeead9